refactor: move background jobs simulation to inside dashboard
